Checkpointing.
Starting to solidify the design of the trie node.  Decided on
a 16 byte structure that gives us everything we need without losing
clock cycles on any sort of bit manipulation of the node data fields
so that we can keep bit manipulation to the key checking part of the
algortihm only.