More updates to prefix handling
We're now not just marking full prefixes, we're also keeping
two indexes in the subnet array to the prefix, both the next most
full one as well as the next sparse one that should be part of the
trie.

This doesn't inflate the mem aligned structures any on amd64, so
we're only paying the cost of the additional clock cycles to parse
through the array yet another time.  I wish there was a way to do
all of this in a single pass, but I don't think the properties of
the data structure will allow for it.  Atleast the time necessary
to build the data structure is only an order of n despite all of
that.