More test tweaks.
Attempting to match for locality of addresses proved to be little
help and actually degraded performance.  This is most likely due
to the subnet being greater than 24 bits in network prefix length.

We're at the point that actually giving this a proper argp input
parser will make sense and allow the test to be significantly more
configurable.

Next up, we need to incorporate bogon list, reading the ASN
descriptions, subnet testing list file, and custom subnets file.
Would help to have a toggle that enables or disables whether or not
reserved subnets are included in the arrays or not.

Still some more interesting work to do, but we're nearing the end.
The algorithm is in place, and this example program just needs some
more spit polish before we can write up a summary.