Performance tuning
Adding max root branch factor of 16, but using the shortest base
subnet length if it's smaller.  This allows us to maximize root
node performance but not compromise the stability of the algorithm
with very short prefixes in the set.

Also, split out the initialization of the private and link local
subnets from the rest of the reserved subnets.  Specials can be
types of bogons if seen on a network, but link local and private
are typically common to see on a network.  Those are bogons on
BGP links, but not inside of a LAN.

Multicast, too, but that and broadcast may need to have their
own initialization routines.